Moved polyfill generation into scripts and fix build process
