+++ package.json
@@ -31,7 +31,8 @@
     "markdown-to-html": "0.0.13",
     "mocha": "4.0.1",
     "performance-now": "2.1.0",
-    "prettier": "1.9.2"
+    "prettier": "1.9.2",
+	"assemblyscript": "github:AssemblyScript/assemblyscript"
   },
   "dependencies": {
     "babel-plugin-mamacro": "0.0.3",
