+++ packages/proposal-sign-extension-ops/src/polyfills/i32_extend8_s.json
@@ -1,0 +1,165 @@
+{
+  "type": "Func",
+  "name": {
+    "type": "Identifier",
+    "value": "packages/proposal-sign-extension-ops/src/polyfills/i32_extend8_s/i32_extend8_s"
+  },
+  "signature": {
+    "type": "Signature",
+    "params": [
+      {
+        "valtype": "i32"
+      }
+    ],
+    "results": [
+      "i32"
+    ]
+  },
+  "body": [
+    {
+      "type": "Instr",
+      "id": "get_local",
+      "args": [
+        {
+          "type": "NumberLiteral",
+          "value": 0,
+          "raw": "0"
+        }
+      ],
+      "loc": {
+        "start": {
+          "line": -1,
+          "column": 58
+        },
+        "end": {
+          "line": -1,
+          "column": 60
+        }
+      }
+    },
+    {
+      "type": "Instr",
+      "id": "const",
+      "args": [
+        {
+          "type": "NumberLiteral",
+          "value": -128,
+          "raw": "-128"
+        }
+      ],
+      "object": "i32"
+    },
+    {
+      "type": "Instr",
+      "id": "or",
+      "args": [],
+      "object": "i32"
+    },
+    {
+      "type": "Instr",
+      "id": "get_local",
+      "args": [
+        {
+          "type": "NumberLiteral",
+          "value": 0,
+          "raw": "0"
+        }
+      ],
+      "loc": {
+        "start": {
+          "line": -1,
+          "column": 64
+        },
+        "end": {
+          "line": -1,
+          "column": 66
+        }
+      }
+    },
+    {
+      "type": "Instr",
+      "id": "const",
+      "args": [
+        {
+          "type": "NumberLiteral",
+          "value": 127,
+          "raw": "127"
+        }
+      ],
+      "object": "i32"
+    },
+    {
+      "type": "Instr",
+      "id": "and",
+      "args": [],
+      "object": "i32"
+    },
+    {
+      "type": "Instr",
+      "id": "get_local",
+      "args": [
+        {
+          "type": "NumberLiteral",
+          "value": 0,
+          "raw": "0"
+        }
+      ],
+      "loc": {
+        "start": {
+          "line": -1,
+          "column": 70
+        },
+        "end": {
+          "line": -1,
+          "column": 72
+        }
+      }
+    },
+    {
+      "type": "Instr",
+      "id": "const",
+      "args": [
+        {
+          "type": "NumberLiteral",
+          "value": 128,
+          "raw": "128"
+        }
+      ],
+      "object": "i32"
+    },
+    {
+      "type": "Instr",
+      "id": "and",
+      "args": [],
+      "object": "i32"
+    },
+    {
+      "type": "Instr",
+      "id": "select",
+      "args": [],
+      "loc": {
+        "start": {
+          "line": -1,
+          "column": 76
+        },
+        "end": {
+          "line": -1,
+          "column": 77
+        }
+      }
+    }
+  ],
+  "loc": {
+    "start": {
+      "line": -1,
+      "column": 56
+    },
+    "end": {
+      "line": -1,
+      "column": 78
+    }
+  },
+  "metadata": {
+    "bodySize": 21
+  }
+}
